requirements.txt
streamlit==1.31.0
python-dotenv==1.0.1
Pillow==10.2.0
huggingface_hub==0.20.3
